Fix CS0266 in GoblinCombatManager gem loot roll
gemLootMin/Max are long, but Random.Range has no (long,long) overload --
the compiler picked Random.Range(float,float), and float doesn't
implicitly convert back to long. Cast through int (gem counts are
discrete anyway) instead. Pre-existing scaffold bug, unrelated to the
art import work.

+++ Assets/_Project/Scripts/Goblins/GoblinCombatManager.cs
@@ -61,7 +61,7 @@ namespace GoldAndGoblins.Goblins
             long gems = 0;
             if (Random.value < currentGoblin.gemDropChance)
             {
-                gems = Random.Range(currentGoblin.gemLootMin, currentGoblin.gemLootMax + 1);
+                gems = Random.Range((int)currentGoblin.gemLootMin, (int)currentGoblin.gemLootMax + 1);
                 CurrencyManager.Instance.AddGems(gems);
             }
